requirements.txt
# Django and Web Framework
Django==5.2.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3

# API and HTTP
requests==2.32.4
aiohttp==3.12.13
websockets==15.0.1

# Data Processing
pandas==2.3.0
numpy==2.2.6
pytz==2025.2

# Financial Data
yfinance==0.2.64
ta==0.11.0

# Authentication and Security
pyotp==2.6.0

# AI/ML Integration
together==1.5.17

# Async and Task Queue
celery==5.3.4
redis==5.0.1

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter (optional)

# Utilities
python-dateutil==2.9.0.post0
colorama==0.4.6
schedule==1.2.0
tabulate==0.9.0

# Development (optional)
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Keep existing packages that are useful
beautifulsoup4==4.13.4
certifi==2025.6.15
click==8.1.8
tqdm==4.67.1
rich==14.0.0
